feat(bpdm-cert): set up docker hub image for application
