+++ src/main/kotlin/org/eclipse/tractusx/bpdmcertificatemanagement/Application.kt
@@ -16,7 +16,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.eclipse.tractusx.bdpmcertificatemanagement
+package org.eclipse.tractusx.bpdmcertificatemanagement
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
