+++ src/main/kotlin/org/eclipse/tractusx/bpdmcertificatemanagement/config/LandingPageConfig.kt
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Apache License, Version 2.0 which is available at
+ * https://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.eclipse.tractusx.bpdmcertificatemanagement.config
+
+import org.springdoc.core.properties.SwaggerUiConfigProperties
+import org.springframework.context.annotation.Configuration
+import mu.KotlinLogging
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry
+
+@Configuration
+class LandingPageConfig(
+    val swaggerProperties: SwaggerUiConfigProperties
+) : WebMvcConfigurer {
+
+    private val logger = KotlinLogging.logger { }
+
+    override fun addViewControllers(registry: ViewControllerRegistry) {
+        val redirectUri = swaggerProperties.path
+        logger.info { "Set landing page to path '$redirectUri'" }
+        registry.addRedirectViewController("/", redirectUri)
+    }
+}
